Minimum set of files needed to  compile code with CMSIS (Version 4.00)
This is an extract from the CMSIS Specification Version 4.00,
downloaded from https://silver.arm.com/download/download.tm?pv=1566150
(Requires free registration)

I have extracted the BSD licenced files that are required to compile
a minimum project.  Not included are the DSP_Lib and RTOS source files.

Also not included are the Driver or Pack files - which additionally are
licenced under the zlib licence